Refactor + Update Finder
Refactor the minions into their own package dndminions.
Update the class getter to include the "dndminions." + class to find the
classes.
