+++ DnDHoardManager/src/dndhoardmanager/GUI.java
@@ -16,6 +16,10 @@ import javax.swing.JOptionPane;
 import javax.swing.JTable;
 import javax.swing.table.DefaultTableModel;
 
+import dndminions.Minions;
+import dndminions.Skele;
+import dndminions.Zombi;
+
 
 /**
  *
@@ -31,7 +35,7 @@ public class GUI extends javax.swing.JFrame {
         
     int wizlv;
     int proficiency; 
-    boolean dynamicallyFindClass = false;
+    boolean dynamicallyFindClass = true;
     
     
     /**
@@ -757,7 +761,7 @@ public class GUI extends javax.swing.JFrame {
     {
         if(dynamicallyFindClass)
         {
-            Class clazz = Class.forName(className);
+            Class clazz = Class.forName("dndminions." + className);
             Minions mini = (Minions) clazz.newInstance();
             return mini;
         }
@@ -766,7 +770,7 @@ public class GUI extends javax.swing.JFrame {
             Minions[] classes = new Minions[]{new Zombi(),new Skele()};
             for (Minions classe : classes)
             {
-                if (classe.name.equals(className))
+                if (classe.getName().equals(className))
                 {
                     return classe;
                 }
