+++ DnDHoardManager/src/dndminions/BrownBear.java
@@ -3,12 +3,12 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package dndhoardmanager;
+package dndminions;
 
 /**
  *
  * @author Ryan
  */
-public class GiantGoat extends Minions{
+public class BrownBear extends Minions{
     
 }
