+++ DnDHoardManager/src/dndminions/KnightStatue.java
@@ -3,12 +3,12 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package dndhoardmanager;
+package dndminions;
 
 /**
  *
  * @author Ryan
  */
-public class FireElemental extends Minions{
+public class KnightStatue extends Minions{
     
 }
