+++ DnDHoardManager/src/dndminions/Minions.java
@@ -3,10 +3,12 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package dndhoardmanager;
+package dndminions;
 
 import javax.swing.JFrame;
 import javax.swing.JOptionPane;
+
+import dndhoardmanager.Attacks;
 
 /**
  *
