+++ DnDHoardManager/src/dndminions/Mummy.java
@@ -3,12 +3,12 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package dndhoardmanager;
+package dndminions;
 
 /**
  *
  * @author Ryan
  */
-public class Goat extends Minions{
+public class Mummy extends Minions{
     
 }
