+++ DnDHoardManager/src/dndminions/Skele.java
@@ -3,8 +3,10 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package dndhoardmanager;
+package dndminions;
 
+import dndhoardmanager.Attacks;
+import dndhoardmanager.Dice;
 
 /**
  *
