+++ DnDHoardManager/src/dndminions/Zombi.java
@@ -3,9 +3,12 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package dndhoardmanager;
+package dndminions;
 
 import java.util.Random;
+
+import dndhoardmanager.Attacks;
+import dndhoardmanager.Dice;
 
 /**
  *
